Refactor code to simplify commands
