+++ Celbridge/Modules/Celbridge.Screenplay/Commands/SaveScreenplayCommand.cs
@@ -1,3 +1,4 @@
+using Celbridge.Activities;
 using Celbridge.Commands;
 using Celbridge.Screenplay.Services;
 using Celbridge.Workspace;
@@ -6,25 +7,18 @@ namespace Celbridge.Screenplay.Commands;
 
 public class SaveScreenplayCommand : CommandBase
 {
-    private readonly IServiceProvider _serviceProvider;
-    private readonly IWorkspaceSettings _workspaceSettings;
+    private readonly IActivityService _activityService;
 
     public ResourceKey WorkbookResource { get; set; } = ResourceKey.Empty;
 
-    public SaveScreenplayCommand(
-        IServiceProvider serviceProvider,
-        IWorkspaceWrapper workspaceWrapper)
+    public SaveScreenplayCommand(IWorkspaceWrapper workspaceWrapper)
     {
-        _serviceProvider = serviceProvider;
-        _workspaceSettings = workspaceWrapper.WorkspaceService.WorkspaceSettings;
+        _activityService = workspaceWrapper.WorkspaceService.ActivityService;
     }
 
     public override async Task<Result> ExecuteAsync()
     {
-        var workspaceWrapper = _serviceProvider.AcquireService<IWorkspaceWrapper>();
-        var activityService = workspaceWrapper.WorkspaceService.ActivityService;
-
-        var getActivityResult = activityService.GetActivity(nameof(ScreenplayActivity));
+        var getActivityResult = _activityService.GetActivity(nameof(ScreenplayActivity));
         if (getActivityResult.IsFailure)
         {
             return Result.Fail($"Failed to get Screenplay activity")
@@ -44,9 +38,6 @@ public class SaveScreenplayCommand : CommandBase
             return Result.Fail($"Failed to save screenplay to workbook")
                 .WithErrors(saveResult);
         }
-
-        // All modified scenes have now been saved, so reset the modified scenes list
-        await _workspaceSettings.DeletePropertyAsync(ScreenplayConstants.ModifiedScenesKey);
 
         return Result.Ok();
     }
